server/requirements.txt
flask
flask-cors
python-dotenv
langchain-core
langchain-openai
langgraph
google-auth
google-auth-oauthlib
google-api-python-client
